requirements.txt
alembic==1.13.1
blinker==1.7.0
bleach==6.2.0
celery==5.4.0
certifi==2024.7.4
click==8.1.7
email-validator==2.3.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flask-security-too==5.6.2
flask-session==0.8.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.5
MarkupSafe==2.1.5
numpy==1.26.4
openai==1.42.0
passlib==1.7.4
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2024.1
pypdf>=4.2.0
pillow>=10.0.0
pillow-heif>=0.16.0
rapidfuzz==3.14.1
redis==6.4.0
spacy==3.5.4
SQLAlchemy==2.0.30
tiktoken==0.11.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
unidecode==1.4.0
Werkzeug==3.0.2
